requirements.txt
elasticsearch==8.11.1
fastapi==0.105.0
hazelcast-python-client==5.3.0
Jinja2==3.1.2
pymongo==4.6.1
python-multipart==0.0.6
uvicorn==0.25.0
pandas
fastapi-sessions
